Sua loi tran so khi xu ly gia xe (int -> long)
- Gia xe o to co the toi vai ty dong, vuot gioi han int (~2,1 ty).
  Truoc day: them xe dat vao gio bi crash (OverflowException) va
  nhan vien khong luu/sua duoc xe gia > 2,1 ty ("Gia xe khong hop le").
- Doi DonGia/ThanhTien trong gio hang va cac tong tien sang kieu long.
- Doi parse gia xe trong Xe_VM sang long.
- Sua thong bao loi dang nhap ghi dung ten DB (DL_OTO).
- Xoa ham tao ma KH khong dung trong DangNhap_VM.

+++ Doan/Doan/ViewModel/User_DichVu_VM.cs
@@ -92,7 +92,7 @@ namespace Doan.ViewModel
                 {
                     MaMatHang = dichVu.MaPT,
                     TenMatHang = dichVu.Ten,
-                    DonGia = (int)(dichVu.Gia ?? 0),
+                    DonGia = (long)(dichVu.Gia ?? 0),
                     SoLuong = 1
                 });
             }
